requirements.dev.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.5
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
ruff==0.3.7
setuptools==58.1.0
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
